improved forum design
